requirements.txt
pandas==2.2.1
requests==2.31.0
streamlit==1.35.0
python-dotenv~=1.0.1
numpy==1.26.4
Pillow==10.2.0
scikit-learn==1.4.0
nltk==3.8.1
faiss-cpu==1.7.4
PyGithub==2.1.1